Handle null digest returned by batchStat
When RemoteOutputService.BatchStat returns a null digest, the remote
output service knows the file exists, but it is incapable of returning
the digest. This may be caused by the kernel holding on to dirty pages
in its write-back cache.

By returning null, ActionMetadataHandler's
fileArtifactValueFromArtifact() will compute the file's digest
separately.

Make sure to run a recent build of bb-clientd to detect these cases.