Properly cancel repository downloads
When cancelling an asynchronous repository download task, an interrupt
signal is sent to the download thread. This doesn't mean that the
download stops immediately. Avoid restarting a download until the
previous download has actually stopped, so that the new download is able
to clean old data without crashing (on Windows).

Fixes #21773